# src/main/resources/openapi.yml
openapi: 3.0.3
info:
  title: Bookstore REST API
  description: API contract for the Spring Boot Bookstore System, matching the ERD and pagination/search features.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local Development Server
security:
  - bearerAuth: []
...
paths:
  /api/users:
    get:
      tags: [Users]
      summary: Get all users with search, filter, and pagination
      operationId: getAllUsers
      parameters:
        - name: keyword
          in: query
          required: false
          description: Filter by username or email (partial, case-insensitive)
          schema:
            type: string
        - name: role
          in: query
          required: false
          description: Filter by role
          schema:
            type: string
            enum: [ADMIN, CUSTOMER]
        - name: enabled
          in: query
          required: false
          description: Filter by account status
          schema:
            type: boolean
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
        - name: sort
          in: query
          required: false
          description: "e.g. username,asc"
          schema:
            type: string
      responses:
        '200':
          description: A list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
    post:
      tags: [Users]
      summary: Register a new user
      operationId: createUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '201':
          description: User registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/users/{userId}:
    parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Users]
      summary: Get a user by ID
      operationId: getUserById
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [Users]
      summary: Fully update a user
      operationId: updateUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found
    patch:
      tags: [Users]
      summary: Partially update a user
      operationId: patchUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPatchRequest'
      responses:
        '200':
          description: User patched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found
    delete:
      tags: [Users]
      summary: Delete a user
      operationId: deleteUser
      responses:
        '204':
          description: User deleted successfully
        '404':
          description: User not found
...
  /api/books/{bookId}/reviews:
    parameters:
      - name: bookId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Reviews]
      summary: Get all reviews for a specific book with pagination
      operationId: getReviewsByBook
      security: []
      parameters:
        - name: userId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: rating
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 5
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
        - name: sort
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: A list of reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReviewResponse'
        '404':
          description: Book not found
    post:
      tags: [Reviews]
      summary: Add a new review to a book
      operationId: createReview
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewRequest'
      responses:
        '201':
          description: Review created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewResponse'
        '400':
          description: Validation error
        '404':
          description: Book or User not found
        '409':
          description: User has already reviewed this book
  /api/reviews/{reviewId}:
    parameters:
      - name: reviewId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      tags: [Reviews]
      summary: Partially update a review (e.g. rating or comment)
      operationId: patchReview
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewPatchRequest'
      responses:
        '200':
          description: Review updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewResponse'
        '404':
          description: Review not found
    delete:
      tags: [Reviews]
      summary: Delete a review
      operationId: deleteReview
      responses:
        '204':
          description: Review deleted successfully
        '404':
          description: Review not found
  /api/categories:
    get:
      tags: [Categories]
      summary: Get all categories with search and pagination
      operationId: getAllCategories
      security: []
      parameters:
        - name: keyword
          in: query
          required: false
          description: Search by name or description
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
        - name: sort
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: A list of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryResponse'
    post:
      tags: [Categories]
      summary: Create a new category
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryRequest'
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/categories/{categoryId}:
    parameters:
      - name: categoryId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Categories]
      summary: Get a category by ID
      operationId: getCategoryById
      security: []
      responses:
        '200':
          description: Category found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '404':
          description: Category not found
    put:
      tags: [Categories]
      summary: Fully update a category
      operationId: updateCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryRequest'
      responses:
        '200':
          description: Category updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '404':
          description: Category not found
    patch:
      tags: [Categories]
      summary: Partially update a category
      operationId: patchCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryPatchRequest'
      responses:
        '200':
          description: Category patched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '404':
          description: Category not found
    delete:
      tags: [Categories]
      summary: Delete a category
      operationId: deleteCategory
      responses:
        '204':
          description: Category deleted
        '404':
          description: Category not found
        '409':
          description: Category still has books assigned and cannot be deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/categories/{categoryId}/books:
    parameters:
      - name: categoryId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Categories]
      summary: Get all books belonging to a category
      operationId: getBooksByCategory
      security: []
      responses:
        '200':
          description: A list of books in the category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookResponse'
        '404':
          description: Category not found
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ReviewRequest:
      type: object
      required: [userId, rating]
      properties:
        userId:
          type: string
          format: uuid
        rating:
          type: integer
          format: int32
          minimum: 1
          maximum: 5
        comment:
          type: string
          maxLength: 1000
    ReviewPatchRequest:
      type: object
      properties:
        rating:
          type: integer
          format: int32
          minimum: 1
          maximum: 5
        comment:
          type: string
          maxLength: 1000
    ReviewResponse:
      type: object
      properties:
        reviewId:
          type: string
          format: uuid
        bookId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        username:
          type: string
        rating:
          type: integer
          format: int32
        comment:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserRequest:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 3
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        address:
          type: string
        role:
          type: string
          enum: [ADMIN, CUSTOMER]
          default: CUSTOMER
        enabled:
          type: boolean
          default: true
    UserPatchRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 3
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        address:
          type: string
        role:
          type: string
          enum: [ADMIN, CUSTOMER]
        enabled:
          type: boolean
    UserResponse:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
        address:
          type: string
        role:
          type: string
          enum: [ADMIN, CUSTOMER]
        enabled:
          type: boolean
...
    BookResponse:
      type: object
      properties:
        bookId:
          type: string
          format: uuid
        title:
          type: string
        author:
          type: string
        price:
          type: number
          format: double
        quantity:
          type: integer
          format: int32
        description:
          type: string
        imgURL:
          type: string
          format: uri
        categoryId:
          type: string
          format: uuid
    CategoryRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 2
        description:
          type: string
    CategoryPatchRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        description:
          type: string
    CategoryResponse:
      type: object
      properties:
        categoryId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
    ErrorItem:
      type: object
      properties:
        code:
          type: string
          example: "E400_INVALID_FIELD"
        message:
          type: string
          example: "must not be null"
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
          example: 400
        error:
          type: string
          example: "Bad Request"
        message:
          type: string
          example: "Validation Failed"
        path:
          type: string
          example: "/api/orders"
        errorItems:
          type: array
          items:
            $ref: '#/components/schemas/ErrorItem'